+++ examples/window_run_return.rs
@@ -10,6 +10,8 @@
 ))]
 fn main() {
     use std::{thread::sleep, time::Duration};
+
+    use simple_logger::SimpleLogger;
     use winit::{
         event::{Event, WindowEvent},
         event_loop::{ControlFlow, EventLoop},
@@ -18,7 +20,7 @@ fn main() {
     };
     let mut event_loop = EventLoop::new();
 
-    simple_logger::init().unwrap();
+    SimpleLogger::new().init().unwrap();
     let _window = WindowBuilder::new()
         .with_title("A fantastic window!")
         .build(&event_loop)