+++ src/platform_impl/android/mod.rs
@@ -44,6 +44,20 @@ fn poll(poll: Poll) -> Option<EventSource> {
 pub struct EventLoop<T: 'static> {
     window_target: event_loop::EventLoopWindowTarget<T>,
     user_queue: Arc<Mutex<VecDeque<T>>>,
+    first_event: Option<EventSource>,
+    start_cause: event::StartCause,
+    looper: ThreadLooper,
+    running: bool,
+}
+
+macro_rules! call_event_handler {
+    ( $event_handler:expr, $window_target:expr, $cf:expr, $event:expr ) => {{
+        if $cf != ControlFlow::Exit {
+            $event_handler($event, $window_target, &mut $cf);
+        } else {
+            $event_handler($event, $window_target, &mut ControlFlow::Exit);
+        }
+    }};
 }
 
 impl<T: 'static> EventLoop<T> {
@@ -56,42 +70,61 @@ impl<T: 'static> EventLoop<T> {
                 _marker: std::marker::PhantomData,
             },
             user_queue: Default::default(),
+            first_event: None,
+            start_cause: event::StartCause::Init,
+            looper: ThreadLooper::for_thread().unwrap(),
+            running: false,
         }
     }
 
-    pub fn run<F>(self, mut event_handler: F) -> !
+    pub fn run<F>(mut self, event_handler: F) -> !
     where
         F: 'static
             + FnMut(event::Event<'_, T>, &event_loop::EventLoopWindowTarget<T>, &mut ControlFlow),
     {
-        let mut cf = ControlFlow::default();
-        let mut first_event = None;
-        let mut start_cause = event::StartCause::Init;
-        let looper = ThreadLooper::for_thread().unwrap();
-        let mut running = false;
+        self.run_return(event_handler);
+        ::std::process::exit(0);
+    }
 
-        loop {
-            event_handler(
-                event::Event::NewEvents(start_cause),
+    pub fn run_return<F>(&mut self, mut event_handler: F)
+    where
+        F: FnMut(event::Event<'_, T>, &event_loop::EventLoopWindowTarget<T>, &mut ControlFlow),
+    {
+        let mut control_flow = ControlFlow::default();
+
+        'event_loop: loop {
+            call_event_handler!(
+                event_handler,
                 self.window_target(),
-                &mut cf,
+                control_flow,
+                event::Event::NewEvents(self.start_cause)
             );
 
             let mut redraw = false;
             let mut resized = false;
 
-            match first_event.take() {
+            match self.first_event.take() {
                 Some(EventSource::Callback) => match ndk_glue::poll_events().unwrap() {
                     Event::WindowCreated => {
-                        event_handler(event::Event::Resumed, self.window_target(), &mut cf);
+                        call_event_handler!(
+                            event_handler,
+                            self.window_target(),
+                            control_flow,
+                            event::Event::Resumed
+                        );
                     }
                     Event::WindowResized => resized = true,
                     Event::WindowRedrawNeeded => redraw = true,
                     Event::WindowDestroyed => {
-                        event_handler(event::Event::Suspended, self.window_target(), &mut cf);
+                        call_event_handler!(
+                            event_handler,
+                            self.window_target(),
+                            control_flow,
+                            event::Event::Suspended
+                        );
                     }
-                    Event::Pause => running = false,
-                    Event::Resume => running = true,
+                    Event::Pause => self.running = false,
+                    Event::Resume => self.running = true,
                     Event::ConfigChanged => {
                         let am = ndk_glue::native_activity().asset_manager();
                         let config = Configuration::from_asset_manager(&am);
@@ -107,7 +140,12 @@ impl<T: 'static> EventLoop<T> {
                                     scale_factor,
                                 },
                             };
-                            event_handler(event, self.window_target(), &mut cf);
+                            call_event_handler!(
+                                event_handler,
+                                self.window_target(),
+                                control_flow,
+                                event
+                            );
                         }
                     }
                     _ => {}
@@ -147,7 +185,12 @@ impl<T: 'static> EventLoop<T> {
                                                     force: None,
                                                 }),
                                             };
-                                            event_handler(event, self.window_target(), &mut cf);
+                                            call_event_handler!(
+                                                event_handler,
+                                                self.window_target(),
+                                                control_flow,
+                                                event
+                                            );
                                         }
                                     }
                                     InputEvent::KeyEvent(_) => {} // TODO
@@ -160,50 +203,69 @@ impl<T: 'static> EventLoop<T> {
                 Some(EventSource::User) => {
                     let mut user_queue = self.user_queue.lock().unwrap();
                     while let Some(event) = user_queue.pop_front() {
-                        event_handler(
-                            event::Event::UserEvent(event),
+                        call_event_handler!(
+                            event_handler,
                             self.window_target(),
-                            &mut cf,
+                            control_flow,
+                            event::Event::UserEvent(event)
                         );
                     }
                 }
                 None => {}
             }
 
-            event_handler(
-                event::Event::MainEventsCleared,
+            call_event_handler!(
+                event_handler,
                 self.window_target(),
-                &mut cf,
+                control_flow,
+                event::Event::MainEventsCleared
             );
 
-            if resized && running {
+            if resized && self.running {
                 let size = MonitorHandle.size();
                 let event = event::Event::WindowEvent {
                     window_id: window::WindowId(WindowId),
                     event: event::WindowEvent::Resized(size),
                 };
-                event_handler(event, self.window_target(), &mut cf);
+                call_event_handler!(event_handler, self.window_target(), control_flow, event);
             }
 
-            if redraw && running {
+            if redraw && self.running {
                 let event = event::Event::RedrawRequested(window::WindowId(WindowId));
-                event_handler(event, self.window_target(), &mut cf);
+                call_event_handler!(event_handler, self.window_target(), control_flow, event);
             }
 
-            event_handler(
-                event::Event::RedrawEventsCleared,
+            call_event_handler!(
+                event_handler,
                 self.window_target(),
-                &mut cf,
+                control_flow,
+                event::Event::RedrawEventsCleared
             );
 
-            match cf {
-                ControlFlow::Exit => panic!(),
+            match control_flow {
+                ControlFlow::Exit => {
+                    self.first_event = poll(
+                        self.looper
+                            .poll_once_timeout(Duration::from_millis(0))
+                            .unwrap(),
+                    );
+                    self.start_cause = event::StartCause::WaitCancelled {
+                        start: Instant::now(),
+                        requested_resume: None,
+                    };
+                    break 'event_loop;
+                }
                 ControlFlow::Poll => {
-                    start_cause = event::StartCause::Poll;
+                    self.first_event = poll(
+                        self.looper
+                            .poll_all_timeout(Duration::from_millis(0))
+                            .unwrap(),
+                    );
+                    self.start_cause = event::StartCause::Poll;
                 }
                 ControlFlow::Wait => {
-                    first_event = poll(looper.poll_all().unwrap());
-                    start_cause = event::StartCause::WaitCancelled {
+                    self.first_event = poll(self.looper.poll_all().unwrap());
+                    self.start_cause = event::StartCause::WaitCancelled {
                         start: Instant::now(),
                         requested_resume: None,
                     }
@@ -215,8 +277,8 @@ impl<T: 'static> EventLoop<T> {
                     } else {
                         instant - start
                     };
-                    first_event = poll(looper.poll_all_timeout(duration).unwrap());
-                    start_cause = if first_event.is_some() {
+                    self.first_event = poll(self.looper.poll_all_timeout(duration).unwrap());
+                    self.start_cause = if self.first_event.is_some() {
                         event::StartCause::WaitCancelled {
                             start,
                             requested_resume: Some(instant),
@@ -234,16 +296,6 @@ impl<T: 'static> EventLoop<T> {
 
     pub fn window_target(&self) -> &event_loop::EventLoopWindowTarget<T> {
         &self.window_target
-    }
-
-    pub fn primary_monitor(&self) -> MonitorHandle {
-        MonitorHandle
-    }
-
-    pub fn available_monitors(&self) -> VecDeque<MonitorHandle> {
-        let mut v = VecDeque::with_capacity(1);
-        v.push_back(self.primary_monitor());
-        v
     }
 
     pub fn create_proxy(&self) -> EventLoopProxy<T> {
@@ -278,6 +330,20 @@ impl<T> Clone for EventLoopProxy<T> {
 
 pub struct EventLoopWindowTarget<T: 'static> {
     _marker: std::marker::PhantomData<T>,
+}
+
+impl<T: 'static> EventLoopWindowTarget<T> {
+    pub fn primary_monitor(&self) -> Option<monitor::MonitorHandle> {
+        Some(monitor::MonitorHandle {
+            inner: MonitorHandle,
+        })
+    }
+
+    pub fn available_monitors(&self) -> VecDeque<MonitorHandle> {
+        let mut v = VecDeque::with_capacity(1);
+        v.push_back(MonitorHandle);
+        v
+    }
 }
 
 #[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
@@ -317,8 +383,10 @@ impl Window {
         WindowId
     }
 
-    pub fn primary_monitor(&self) -> MonitorHandle {
-        MonitorHandle
+    pub fn primary_monitor(&self) -> Option<monitor::MonitorHandle> {
+        Some(monitor::MonitorHandle {
+            inner: MonitorHandle,
+        })
     }
 
     pub fn available_monitors(&self) -> VecDeque<MonitorHandle> {
@@ -327,10 +395,10 @@ impl Window {
         v
     }
 
-    pub fn current_monitor(&self) -> monitor::MonitorHandle {
-        monitor::MonitorHandle {
+    pub fn current_monitor(&self) -> Option<monitor::MonitorHandle> {
+        Some(monitor::MonitorHandle {
             inner: MonitorHandle,
-        }
+        })
     }
 
     pub fn scale_factor(&self) -> f64 {
@@ -358,7 +426,7 @@ impl Window {
     }
 
     pub fn set_inner_size(&self, _size: Size) {
-        panic!("Cannot set window size on Android");
+        warn!("Cannot set window size on Android");
     }
 
     pub fn outer_size(&self) -> PhysicalSize<u32> {
@@ -380,7 +448,7 @@ impl Window {
     pub fn set_maximized(&self, _maximized: bool) {}
 
     pub fn set_fullscreen(&self, _monitor: Option<window::Fullscreen>) {
-        panic!("Cannot set fullscreen on Android");
+        warn!("Cannot set fullscreen on Android");
     }
 
     pub fn fullscreen(&self) -> Option<window::Fullscreen> {